added preempting support to zmqclient; re-organized exceptions
